# shared/openapi.yaml
openapi: 3.1.0
info:
  title: cinghialapp API
  version: 0.1.0
  description: |
    D&D 5e campaign manager API.
    Server-authoritative dice. Role-based visibility (player/master).
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterReq' }
      responses:
        '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthRes' } } } }
  /auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginReq' }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/AuthRes' } } } }
...
  /campaigns:
    get:
      tags: [campaigns]
      responses: { '200': { description: list } }
    post:
      tags: [campaigns]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CampaignCreate' } } }
      responses: { '201': { description: created } }
  /campaigns/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:    { tags: [campaigns], responses: { '200': { description: ok } } }
    patch:  { tags: [campaigns], responses: { '200': { description: ok } } }
    delete: { tags: [campaigns], responses: { '204': { description: deleted } } }
  /campaigns/{id}/members:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:  { tags: [campaigns], responses: { '200': { description: ok } } }
    post: { tags: [campaigns], responses: { '201': { description: added } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
    User:
      type: object
      required: [id, email, display_name, role, language]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [user, admin] }
        language: { type: string, enum: [en, it] }
        avatar_url: { type: string, nullable: true }
    RegisterReq:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        display_name: { type: string, minLength: 1, maxLength: 64 }
        language: { type: string, enum: [en, it], default: en }
    LoginReq:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthRes:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:  { $ref: '#/components/schemas/User' }
    CampaignCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
...